backend/requirements.txt
python-dotenv
aioboto3
cryptography
uvicorn
fastapi
PyJWT
